Rename flowctl test to check
It's convenient to use this subcommand to trigger the generation of
Typescript source code for derivations. This has been documented as part
of the workflow. Rather than add a new subcommand for source generation,
this renames the existing test subcommand in an attempt to free it from
an expectation that it will _only_ run tests.